// src/app/modules/globalSetting/globalSetting.interface.ts
export interface IGlobalSetting {
  name: string;
  description: string;
  businessNumber: string;
  businessAddress: string;
  businessLocation: string;
  businessSlogan: string;
  businessEmail: string;
  businessFacebook: string;
  businessInstagram: string;
  businessTwitter: string;
  businessLinkedin: string;
  businessYoutube: string;
  businessWhatsapp: string;
  businessWorkHours: string;
  deliveryChargeInsideDhaka: string;
  deliveryChargeOutsideDhaka: string;
  primaryColor: string;
  secondaryColor: string;
  deliveryApiKey: string;
  deliverySecretKey: string;
  logo: string;
  favicon: string;
  currency: string;
  ssl: string;
  status: string;
}